Revert "sm8250-common: Build libsecril-client from sources"
The source built libsec ril-client is not accessible from the vendor because it is located in the system partition.

CANNOT LINK EXECUTABLE "/vendor/bin/hw/android.hardware.secure_element@1.0-service":
library "libsecril-client.so" not found: needed by /vendor/lib64/libsec_semRil.so in namespace (default)

It's useless in this state.

This reverts commit 5dda39e0d537ed16c2ad85d9422156c8c115af4a.

Change-Id: Id26098df91a5aa3a7edaa1cbaf730297a729ea03